requirements.txt
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.4
Flask==2.2.3
Flask-DotEnv==0.1.2
flask-swagger-ui==4.11.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycodestyle==2.10.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.29.0
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
